requirements.txt
# Ising Life Lab - Python Dependencies
# Computational core
numpy>=1.24.0
scipy>=1.10.0

# Configuration management
pyyaml>=6.0
jsonschema>=4.17.0

# Data handling
pandas>=2.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Reproducibility
python-dotenv>=1.0.0

